Some bugs fixed

+++ Src/WebAPI/Controllers/ConditionsOfProductController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using WebAPI.DTOs.ConditionsOfProductDtos;
 
@@ -12,6 +13,7 @@ namespace WebAPI.Controllers
 {
     [Route("api/v1/Conditions")]
     [ApiController]
+    [Authorize]
     public class ConditionsOfProductController : ControllerBase
     {
         private readonly IUnitOfWork _unitOfWork;
@@ -22,35 +24,42 @@ namespace WebAPI.Controllers
         }
 
         [HttpGet]
-        //[Authorize]
         public async Task<IActionResult> Get()
         {
             return Ok(await _unitOfWork.ConditionsOfProductService.GetAllConditions());
         }
 
         [HttpPost]
-        //[Authorize(Roles = "admin")]
         public async Task<IActionResult> CreateCondition([FromBody] CreateConditionDto conditionDto)
         {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest("Invalid data.");
-            }
 
-            await _unitOfWork.ConditionsOfProductService.AddCondition(conditionDto.Condition);
-            _unitOfWork.Complete();
-            return Ok();
+            string Roletype = User.FindFirstValue("Roletype");
+            if (Roletype == "False")
+            {
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest("Invalid data.");
+                }
+
+                await _unitOfWork.ConditionsOfProductService.AddCondition(conditionDto.Condition);
+                _unitOfWork.Complete();
+                return Ok();
+            }
+            return Unauthorized();
         }
 
         [HttpDelete("{id}")]
-        [Authorize(Roles = "admin")]
         public async Task<IActionResult> DeleteCondition(int id)
         {
-            await _unitOfWork.ConditionsOfProductService.DeleteCondition(id);
-            _unitOfWork.Complete();
-            return NoContent();
+            string Roletype = User.FindFirstValue("Roletype");
+            if (Roletype == "False")
+            {
+                await _unitOfWork.ConditionsOfProductService.DeleteCondition(id);
+                _unitOfWork.Complete();
+                return NoContent();
+            }
+            return Unauthorized();
         }
-
 
     }
 }
